style: rename cve source crate
