+++ domain-db/src/lib.rs
@@ -4,4 +4,4 @@ extern crate r2d2;
 extern crate r2d2_diesel;
 
 pub mod db;
-pub mod sources;
+pub mod cve_sources;
